+++ NEWS.md
@@ -1,3 +1,9 @@
+# VICmodel 0.1.2
+
+* Add some function to calculate Nash-Sutcliffe coefficient of efficiency (NSE) and relative bias (RB) to evaluate the streamflow simulation performance.
+
+* Longwave radiation and atmospheric pressure are not necessary forcing input now.
+
 # VICmodel 0.1.1
 
 * Fix some codes with bug or potential problem of the VIC source codes.